Change use statements and use of usize for best practices
